move some class in the package they belong to
